FIX: (#145) 인증도메인 인프라스트럭처 계층을 재정의한다
