--- src/main/java/com/zerozero/auth/infrastructure/oauth/common/Provider.java
@@ -1,5 +1,0 @@
-package com.zerozero.auth.infrastructure.oauth.common;
-
-public enum Provider {
-  KAKAO
-}
