+++ src/main/java/com/zerozero/auth/infrastructure/oauth/core/Provider.java
@@ -1,0 +1,5 @@
+package com.zerozero.auth.infrastructure.oauth.core;
+
+public enum Provider {
+    KAKAO
+}
